requirements.txt
torch
transformers
datasets
streamlit
pdfminer.six
python-docx
rouge-score